Fix vesu context serialization for v1 and v2
