+++ packages/nextjs/components/modals/stark/WithdrawModalStark.tsx
@@ -1,7 +1,8 @@
 import { FC } from "react";
 import { TokenActionModal, TokenInfo } from "../TokenActionModal";
 import { formatUnits } from "viem";
-import { VesuContext, useLendingAction } from "~~/hooks/useLendingAction";
+import { useLendingAction } from "~~/hooks/useLendingAction";
+import type { VesuContext } from "~~/utils/vesu";
 import { PositionManager } from "~~/utils/position";
 
 interface WithdrawModalStarkProps {
